Add supoort do map.

+++ src/main/java/com/hdi/ruralbuscabff/api/model/dto/ProdutoDto.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ProdutoDto {
 
-    private Integer produtoId;
+    private Long produtoId;
     private String descricaoProduto;
 }
